% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmse.R
\name{pmse}
\alias{pmse}
\title{Adaptation of PMSE function from singR Package}
\usage{
pmse(M1 = NULL, M2 = NULL, S1 = NULL, S2 = NULL, standardize = FALSE)
}
\arguments{
\item{M1}{A numeric matrix (d x p) representing an estimated structure.}

\item{M2}{A numeric matrix (q x p) representing a reference structure.}

\item{S1}{A numeric matrix (n x p) representing estimated subject scores.}

\item{S2}{A numeric matrix (n x q) representing reference subject scores.}

\item{standardize}{Logical (TRUE/FALSE): If TRUE, standardizes matrices to unit variance before computing PMSE (default = FALSE).}
}
\value{
A numeric value. The Procrustes mean squared error.
}
\description{
Adaptation of PMSE function from singR Package
}
\examples{
M1 <- matrix(rnorm(20), 5, 4)
M2 <- matrix(rnorm(20), 5, 4)
pmse(M1 = M1, M2 = M2)

S1 <- matrix(rnorm(30), 10, 3)
S2 <- matrix(rnorm(30), 10, 3)
pmse(S1 = S1, S2 = S2)
}
