% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProJIVE.R
\name{ProJIVE}
\alias{ProJIVE}
\title{Wrapper function to conduct ProJIVE analyses (w option for multiple initial values for loadings), and calculate asymptotic variance}
\usage{
ProJIVE(
  Y,
  P,
  Q,
  Max.iter = 5000,
  diff.tol = 1e-05,
  plots = TRUE,
  chord.tol = -1,
  sig_hat = NULL,
  init.loads = NULL,
  center = FALSE,
  num.starts = 5,
  AsymVar = FALSE,
  return.all.starts = FALSE
)
}
\arguments{
\item{Y}{Matrix. The observed data.}

\item{P}{Integer vector. The number of variables in each dataset.}

\item{Q}{Integer vector. The number of components in each dataset.}

\item{Max.iter}{The maximum number of iterations for the EM algorithm (default = 5000).}

\item{diff.tol}{Numeric. Specifies the tolerance for convergence based on log-likelihood (default = 1e-5).}

\item{plots}{Logical (TRUE/FALSE): Specifies if plots are generated (default = TRUE).}

\item{chord.tol}{Numeric. Tolerance for the chordal distance for convergence (default = -1).}

\item{sig_hat}{The error variances for each dataset.}

\item{init.loads}{Either a list of initial loadings or a string specifying the initialization method ("AJIVE" or "CJIVE").}

\item{center}{Logical (TRUE/FALSE): Specifies if the data matrix is centered before estimation (default = FALSE).}

\item{num.starts}{Integer. The number of different initializations (default = 5).}

\item{AsymVar}{Logical (TRUE/FALSE): If TRUE, computes the asymptotic variance using the observed empirical information matrix (default = FALSE).}

\item{return.all.starts}{Logical (TRUE/FALSE): If TRUE, returns results from all initializations; if FALSE, return only the best result (default: FALSE).}
}
\value{
A list containing:
- ProJIVE_Results: The best ProJIVE estimation result (or all results if "return.all.starts = TRUE").
- ObservedDataLogLikelihood: The observed log-likelihood for each initialization.
- ObservedEmpiricalInfo (when "AsymVar = TRUE"): The observed empirical information matrix for asymptotic variance estimation.
}
\description{
Wrapper function to conduct ProJIVE analyses (w option for multiple initial values for loadings), and calculate asymptotic variance
}
\examples{
Y <- matrix(rnorm(100), 20, 5)
P <- c(5, 5)
Q <- c(2, 3)
result <- ProJIVE(Y, P, Q)
str(result)
}
