% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProJIVE_AsymVar.R
\name{ProJIVE_AsymVar}
\alias{ProJIVE_AsymVar}
\title{Calculate empirical observed information matrix}
\usage{
ProJIVE_AsymVar(W.mats, error.vars, theta, r.J, Y)
}
\arguments{
\item{W.mats}{A list of matrices that each represents the variable loadings of the dataset. Rows as variables and columns as latent components.}

\item{error.vars}{Variance of the noise for each dataset.}

\item{theta}{A matrix that represents the subject scores.}

\item{r.J}{Integer. The number of joint components in the PJIVE decomposition.}

\item{Y}{A matrix that represents the observed data. Rows as subjects and columns as features from all datasets combined.}
}
\value{
A list containing:
- MeanScoreVector: The mean of the score vectors across subjects.
- ObservedEmpiricalInformationMatrix: The empirical observed information matrix.
- Inverse_ObservedEmpiricalInformationMatrix: The inverse of the observed information matrix.
}
\description{
Calculate empirical observed information matrix
}
\examples{
W1 <- matrix(rnorm(50), 10, 5)
W2 <- matrix(rnorm(60), 12, 5)
W.mats <- list(W1, W2)
error.vars <- c(0.5, 0.8)
theta <- matrix(rnorm(100), 20, 5)
r.J <- 2
Y <- matrix(rnorm(400), 20, 20)
result <- ProJIVE_AsymVar(W.mats, error.vars, theta, r.J, Y)
str(result)
}
